Support projections-style transformations to reduce false-positives
Basic aliases (e.g. admin?/be_admin) can be represented easily with simple
wildcards, but more complex transformations require a different mechanism.

Instead of using %s to represent strings that can be replaced 1:1, this
introduces a syntax inspired by https://github.com/tpope/vim-projectionist, as
such:

    - name: Rails
      aliases:
      - from: "*Validator"
        to: "{snakecase}"

This would find `AbsoluteUriValidator` and also match `absolute_uri`, which
would be found if the validation was in use.

This currently supports the `camelcase` and `snakecase` transformations.

Closes #18